third_party/blink/web_tests/fast/block/margin-collapse/display-contents-contenteditable.html

<!DOCTYPE html>
<p>There should be a green square below.</p>
<div id="container" style="overflow:hidden; width:100px; background:green;" data-expected-height="100">
  <div style="display:contents;">
    <div id="initiallyNone" style="display:none; margin:100px 0;">
      <div style="margin:100px 0;"></div>
    </div>
  </div>
  <div style="margin:100px 0;">
    <div style="margin:100px 0;"></div>
  </div>
  <div contenteditable="true" style="height:0;"></div>
</div>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/check-layout-th.js"></script>
<script>
  document.body.offsetTop;
  document.getElementById("initiallyNone").style.display = "block";
  checkLayout("#container");
</script>
